fix:修改user資料庫欄位
