+++ src/routes/loginRoutes.js
@@ -3,6 +3,5 @@ const router = express.Router();
 const { login } = require("../controllers/authController");
 
 router.post("/login", login);
-router.put("/:id/password", updateUserPassword);
 
 module.exports = router;
