--- src/routes/memberRoutes.js
@@ -1,14 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const upload = require("../middlewares/uploadAvatar");
-const {
-    getMember,
-    updateMember,
-    uploadAvatar
-} = require('../controllers/memberController');
-
-router.get('/:id', getMember);
-router.put('/:id', updateMember);
-router.post('/upload-avatar', upload.single('avatar'), uploadAvatar);
-
-module.exports = router;
